pyproject.toml
[tool.poetry]
name = "pathway-agentic-rag"
version = "0.1.0"
description = ""
authors = ["Karthik Ayangar <karthik.ayangar7118@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
docling = "^2.4.1"
requests = "^2.32.3"
swarm = {git = "https://github.com/openai/swarm.git"}
duckduckgo-search = "^6.3.3"
tavily-python = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
